🔍 refactor. element 정리
