+++ src/containers/header/Header.tsx
@@ -4,11 +4,11 @@ import { Link } from "react-router-dom";
 import { openDB } from "idb";
 import { useAppDispatch, useAppSelector } from "../../redux/store";
 import { setCartData } from "../../redux/reducer/cartSlice";
-import { SearchBar } from "./../../elements/SearchBar";
+import { SearchBar } from "../../elements/searchBar/SearchBar";
 import headLogo from "./../../assets/logo/headLogo.png";
 import SchProdModal from "../../components/modal/schProdModal/SchProdModal";
 import ScrollHeader from "../../components/header/scroll/ScrollHeader";
-import MenuDrop from "../../elements/MenuDrop";
+import MenuDrop from "../../elements/menuDrop/MenuDrop";
 import useDropDown from "../../hooks/useDropDown";
 import AsideHeader from "../../components/header/aside/AsideHeader";
 import { useLogOut } from "./useLogout";
